requirements.txt
BeautifulSoup==3.2.1
Cython==0.17.1
Django==1.3
execnet==1.1
Fabric==1.5.1
Pillow==1.7.8
Pinax==0.9a2
South==0.7.6
ajaxuploader==0.2.3
argparse==1.2.1
boto==2.6.0
cov-core==1.7
coverage==3.6
dj-database-url==0.2.1
django-announcements==0.1.2
django-annoying==0.7.6
django-appconf==0.5
django-braces==0.2.1
-e git+https://github.com/jezdez/django_compressor.git@c46e52a20c83a1aa552490266c8b2fa942828211#egg=django_compressor-dev
django-countries==1.3
django-crispy-forms==1.2.0
django-dbsettings==0.5
django-debug-toolbar==0.9.4
django-email-confirmation==0.2
django-facebook==4.2.2
django-mailer==0.1.0
django-notification==0.2
git+https://github.com/simonw/django-openid.git#egg=django-openid
django-pagination==1.0.7
django-pylibmc-sasl==0.2.4
django-social-auth==0.7.17
django-staticfiles==1.0.1
django-storages==1.1.5
django-timezones==0.2
django-tinymce==1.5.1b2
django-tracking==0.3.5
django-uni-form==0.7.0
easy-thumbnails==1.1
facebook-sdk==0.3.2
gdata==2.0.17
gevent==0.13.8
gevent-socketio==0.3.5-rc2
gevent-websocket==0.3.6
greenlet==0.4.0
gunicorn==0.15.0
httplib2==0.7.6
ipython==0.13
lxml==3.0.1
oauth2==1.5.211
paramiko==1.9.0
py==1.4.12
pycrypto==2.6
-e git+https://github.com/lericson/pylibmc.git#egg=pylibmc
pytest==2.3.4
pytest-cov==1.6
pytest-xdist==1.8
python-cjson==1.0.5
python-cloudfiles==1.7.11
python-memcached==1.48
python-openid==2.2.5
python-swiftclient==2.0.3
pytz==2014.2
redis==2.7.1
simplejson==2.6.2
wsgiref==0.1.2
tweepy==1.13
MySQL-python==1.2.5
rackspace-auth-openstack==1.3
rackspace-novaclient==1.4
rax-default-network-flags-python-novaclient-ext==0.2.4
rax-scheduled-images-python-novaclient-ext==0.2.1
pymongo==2.7

